+++ CHANGELOG.md
@@ -3,14 +3,19 @@ All notable changes to this project will be documented in this file.
 
 See [DocuSign Support Center](https://support.docusign.com/en/releasenotes/) for Product Release Notes.
 
-## [v3.23.0] - eSignature API v2.1-23.2.00.00 - 2023-05-15
+## [v3.24.0] - eSignature API v2.1-23.3.00.00 - 2023-08-21
+### Changed
+- Added support for version v2.1-23.3.00.00 of the DocuSign ESignature API.
+- Updated the SDK release version.
+
+## [v3.23.0] - eSignature API v2.1-23.2.00.00 - 2023-05-18
 ### Changed
 - Added support for version v2.1-23.2.00.00 of the DocuSign ESignature API.
 - Updated the SDK release version.
 
-## [v3.22.0] - eSignature API v2.1-23.1.01.01 - 2023-04-05
+## [v3.22.0] - eSignature API v2.1-23.2.00.00 - 2023-05-17
 ### Changed
-- Added support for version v2.1-23.1.01.01 of the DocuSign ESignature API.
+- Added support for version v2.1-23.2.00.00 of the DocuSign ESignature API.
 - Updated the SDK release version.
 
 ## [v3.21.0] - eSignature API v2.1-23.1.01.00 - 2023-03-17